backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
pydantic-settings==2.6.1
email-validator==2.2.0
alembic==1.13.2
